refactor: organizse instructions
