+++ libs/angular/internal/src/public_api.ts
@@ -1,3 +1,3 @@
 export { GetManyOptions } from './get-many-options.interface';
 export { OperationContext } from './operation-context.enum';
-
+export { RouteInstructionType } from './route-instruction-type.enum';
