+++ libs/angular/meta/src/meta-resolver/constants.ts
@@ -1,0 +1,1 @@
+export const META_KEY = 'META';
