+++ libs/angular/meta/src/meta-resolver/helper.ts
@@ -1,0 +1,9 @@
+import { RouteInstructionType } from '@lyxs/angular/internal';
+import { RouteInstruction } from '@lyxs/angular/routing';
+
+export function extractMeta(routeDataArray: RouteInstruction[]) {
+    const meta = routeDataArray
+        .find(data => data.type === RouteInstructionType.Meta);
+
+    return meta;
+}
