+++ libs/angular/meta/src/public_api.ts
@@ -1,0 +1,2 @@
+export { MetaResolver } from './meta-resolver/meta.resolver';
+
