--- libs/angular/routing/src/crud-route-data/constants.ts
@@ -1,2 +1,0 @@
-export const CRUD_ROUTE_DATA = 'CRUD_ROUTE_DATA';
-export const CRUD_ROUTE_POPULATE = 'CRUD_ROUTE_POPULATE';
