+++ libs/angular/routing/src/public_api.ts
@@ -1,4 +1,14 @@
-export { CRUD_ROUTE_DATA } from './crud-route-data/constants';
-export { deactivate, getActive, getMany, getOne, instructions, populate, PopulationOptions, PopulationStrategy, reset } from './crud-route-data/route-data';
-export { lyxsRoutes } from './crud-route-resolver/crud-routes';
+export { ROUTING_INSTRUCTION } from './routing-instructions/constants';
+export { instructions } from './routing-instructions/instructions-wrapper';
+export { deactivate } from './routing-instructions/instructions/deactivate.instruction';
+export { getActive } from './routing-instructions/instructions/get-active.instruction';
+export { getMany } from './routing-instructions/instructions/get-many.instruction';
+export { getOne } from './routing-instructions/instructions/get-one.instruction';
+export { meta } from './routing-instructions/instructions/meta.instruction';
+export { populate } from './routing-instructions/instructions/populate.instruction';
+export { PopulationOptions } from './routing-instructions/instructions/population-options.type';
+export { PopulationStrategy } from './routing-instructions/instructions/population-strategy.enum';
+export { reset } from './routing-instructions/instructions/reset.instruction';
+export { RouteInstruction } from './routing-instructions/instructions/route-instruction.type';
+export { lyxsRoutes } from './routing-resolver/lyxs-routes';
