+++ libs/angular/routing/src/routing-instructions/constants.ts
@@ -1,0 +1,1 @@
+export const ROUTING_INSTRUCTION = 'ROUTING_INSTRUCTION';
