+++ libs/angular/routing/src/routing-instructions/instructions-wrapper.ts
@@ -1,0 +1,8 @@
+import { ROUTING_INSTRUCTION } from './constants';
+import { RouteInstruction } from './instructions/route-instruction.type';
+
+export function instructions(data: { [key: string]: RouteInstruction | Array<RouteInstruction> }) {
+    return {
+        [ROUTING_INSTRUCTION]: data,
+    };
+}
