+++ libs/angular/routing/src/routing-instructions/instructions/deactivate.instruction.ts
@@ -1,0 +1,7 @@
+import { RouteInstructionType } from '@lyxs/angular/internal';
+
+export function deactivate() {
+    return {
+        type: RouteInstructionType.Deactivate as RouteInstructionType.Deactivate,
+    };
+}
