+++ libs/angular/routing/src/routing-instructions/instructions/get-active.instruction.ts
@@ -1,0 +1,18 @@
+import { RouteInstructionType } from '@lyxs/angular/internal';
+import { GetOneInstruction } from './get-one.instruction';
+import { instructionDefaults } from './instruction-defaults';
+
+export function getActive(
+    options: GetOneInstruction = {},
+) {
+    options = {
+        ...instructionDefaults,
+        param: 'id',
+        ...options,
+    };
+
+    return {
+        type: RouteInstructionType.GetActive as RouteInstructionType.GetActive,
+        ...options,
+    };
+}
