+++ libs/angular/routing/src/routing-instructions/instructions/get-one.instruction.ts
@@ -1,0 +1,22 @@
+import { RouteInstructionType } from '@lyxs/angular/internal';
+import { Instruction, instructionDefaults } from './instruction-defaults';
+
+export interface GetOneInstruction extends Instruction {
+    /** Key in route.params containing the id to resolve (default: `'id'`) */
+    param?: string;
+}
+
+export function getOne(
+    options: GetOneInstruction = {},
+) {
+    options = {
+        ...instructionDefaults,
+        param: 'id',
+        ...options,
+    };
+
+    return {
+        type: RouteInstructionType.GetOne as RouteInstructionType.GetOne,
+        ...options,
+    };
+}
