+++ libs/angular/routing/src/routing-instructions/instructions/instruction-defaults.ts
@@ -1,0 +1,14 @@
+export interface Instruction {
+    /** Await until loaded (default: `false`) */
+    await?: boolean;
+    /**
+     * During server side rendering, you probably want to wait for your resource to load
+     * before resolving, regardless of the browser behaviour specified by `await`.
+     */
+    awaitPlatformServer?: boolean;
+}
+
+export const instructionDefaults = {
+    await: false,
+    awaitPlatformServer: true,
+} as Instruction;
