+++ libs/angular/routing/src/routing-instructions/instructions/meta.instruction.ts
@@ -1,0 +1,16 @@
+import { RouteInstructionType } from '@lyxs/angular/internal';
+
+export interface MetaInstruction {
+    title?: string;
+    description?: string;
+    image?: string;
+}
+
+export function meta(
+    options: MetaInstruction = {},
+) {
+    return {
+        type: RouteInstructionType.Meta as RouteInstructionType.Meta,
+        ...options,
+    };
+}
