+++ libs/angular/routing/src/routing-instructions/instructions/population-options.type.ts
@@ -1,0 +1,3 @@
+import { populate } from './populate.instruction';
+
+export type PopulationOptions = ReturnType<typeof populate>;
