+++ libs/angular/routing/src/routing-instructions/instructions/reset.instruction.ts
@@ -1,0 +1,7 @@
+import { RouteInstructionType } from '@lyxs/angular/internal';
+
+export function reset() {
+    return {
+        type: RouteInstructionType.Reset as RouteInstructionType.Reset,
+    };
+}
