+++ libs/angular/routing/src/routing-resolver/constants.ts
@@ -1,0 +1,1 @@
+export const RESOLVED_ROUTING_INSTRUCTION_KEY = 'RESOLVED_ROUTING_INSTRUCTION';
