+++ tsconfig.json
@@ -21,7 +21,8 @@
       "@lyxs/angular/ui": ["libs/angular/ui/src/public_api.ts"],
       "@lyxs/angular/pagination": ["libs/angular/pagination/src/public_api.ts"],
       "@lyxs/angular/registry": ["libs/angular/registry/src/public_api.ts"],
-      "@lyxs/angular/internal": ["libs/angular/internal/src/public_api.ts"]
+      "@lyxs/angular/internal": ["libs/angular/internal/src/public_api.ts"],
+      "@lyxs/angular/meta": ["libs/angular/meta/src/public_api.ts"]
     }
   },
   "exclude": ["node_modules", "tmp"]
